tests: Add a final report for the test batch
In the new approach for tests execution, `gtest` is called in a loop with
one test at a time. Therefore, there is no final report from `gtest`
summarizing the results. Here that report is generated from bash.
Also, the job is reported to fail if at least one test failed.

Change-Id: Ica0ee2179f3339c448cb89c7e6e9c2b0c5c4377d