tests: Add optional parallel-tests-dispatcher service
The parallel execution of the tests has been using a partitioning
algorithm to distribute the tests between the workers. This algorithm
depends on the expected time for the tests and the quantity
of available workers. It requires keeping those values updated and
could be error-prone if the deployment conditions change in the testing
infrastructure.

This commit adds a new approach using a web service for dispatching the
tests to the different workers. It doesn't rely on the number of workers
nor on the expected time. However, best performance should be
achieved if the provided list of tests is sorted in descending order
on the expected time.

The change is backward compatible: the previous algorithm continues
working as it was, until two new arguments are added to the
`run_test_concurrently.sh` call, namely: BUILD_ID, and DISPATCHER_URL.
When these arguments are present, the dispatcher's client is used
instead of the previous partitioning algorithm.

The dispatcher uses Flask for the web service. It also has a Dockerfile
and a docker-compose.yml configuration file for easing the deployment.

Last change in this commit:

tests: Set HTTP response status to success on empty data structure

The server cleans the data structure when it gets empty. Then, when a
different worker asks for the same build_id and test_suite, it should
not receive an error, rather the same result as when the list is empty.

Change-Id: Ia3dfc76fd2ab9c0d552d81cae97c0aa7cfe8120d