utils: Add sequence diagram generator for the protocol
When there is the need to explain the LizardFS protocol it gets
complex and extensive. It would be appropriate to have a graphical
representation for a better understanding. Sequence diagrams are
a common tool for describing interactions between elements of
the system architecture.

The new script parses the `MFSCommunication.h` file and produces
a declarative description that can be fed to a sequence diagram
generator, e.g., https://sequencediagram.org/

Change-Id: Iabbd6c9eb5344581375c3a26765b4ae16792c050