tests: Add client script for the tests dispatcher service
After creating the parallel-tests-dispatcher service, it is needed
to modify the `run_test_concurrently.sh` script for using it. A
helper client allows the script to push the required tests to the
service and then iterate concurrently over that list, processing
the next available test until the list is depleted.

The change is backward compatible: the previous algorithm continues
working as it was, until two new arguments are added to the
`run_test_concurrently.sh` call, namely: BUILD_ID, and DISPATCHER_URL.
When these arguments are present, the dispatcher's client is used
instead of the previous partitioning algorithm.

It also includes a new LIZARDFS_TESTS_PATH argument that is expected
to point to the `lizardfs-tests` binary location.

Co-authored-by: guillex <guillex@lizardfs.org>
Change-Id: I996d37659ba198f2b3b7a601663941100e40c8a5